Move factory-less helpers out of pi's scanned extensions root
pi loads every extensions/*.ts and extensions/*/index.ts as an entry point
and refuses to start when one does not default-export a factory. The shared
helpers output-guard.ts, web-transport.ts and project-approval.ts export
none, so pi exited 1 on boot with every release since they were added,
while the unit gate stayed green. Move them to extensions/internal/, which
the loader does not scan, and pin the loader contract with a test that
imports every scanned entry and asserts the factory export.
