+++ extensions/internal/output-guard.ts
